Hopefully fixed compilation? (#371)
* Hopefully fixed compilation?

* fixed version ts
