+++ modules/components/tag/tag.component.ts
@@ -236,7 +236,7 @@ export class TagComponent {
     private disableEditMode($event: KeyboardEvent): void {
         this.editModeActivated = false;
         $event.preventDefault();
-        // http://stackoverflow.com/questions/39787038/how-to-manage-angular2-expression-has-changed-after-it-was-checked-exception-w
+
         this.cdRef.detectChanges();
     }
