--- modules/ng2-tag-input.module.ts
@@ -1,45 +1,0 @@
-import { NgModule } from '@angular/core';
-import { ReactiveFormsModule, FormsModule } from '@angular/forms';
-import { CommonModule } from '@angular/common';
-
-import { Ng2DropdownModule } from 'ng2-material-dropdown';
-
-import {
-    TagInputComponent,
-    TagInputForm,
-    DeleteIconComponent,
-    TagInputDropdown,
-    TagComponent,
-    TagRipple
-} from './components';
-
-import { HighlightPipe } from './core';
-
-@NgModule({
-    imports: [
-        CommonModule,
-        ReactiveFormsModule,
-        FormsModule,
-        Ng2DropdownModule
-    ],
-    declarations: [
-        TagInputComponent,
-        DeleteIconComponent,
-        TagInputForm,
-        TagComponent,
-        HighlightPipe,
-        TagInputDropdown,
-        TagRipple
-    ],
-    exports: [
-        TagInputComponent,
-        DeleteIconComponent,
-        TagInputForm,
-        TagComponent,
-        HighlightPipe,
-        TagInputDropdown,
-        TagRipple
-    ]
-})
-export class TagInputModule {}
-
