+++ package.json
@@ -1,14 +1,15 @@
 {
   "name": "ng2-tag-input",
-  "version": "1.2.1-beta.1",
+  "version": "1.2.1",
   "description": "Tag Input component for Angular 2",
   "scripts": {
     "prepublish": "npm run clean && npm run build",
-    "build": "webpack && npm run moveassets && npm run ngc",
+    "build": "webpack && npm run tsc && npm run moveassets && npm run ngc",
     "watch": "webpack -- --watch",
     "server": "webpack-dev-server --config webpack.demo.js --inline --colors --progress --port 3010 --content-base demo",
     "start": "npm run server",
     "test": "karma start karma.conf.js",
+    "tsc": "tsc -p tsconfig.json",
     "ngc": "ngc -p ./tsconfig.aot.json",
     "moveassets": "rsync -a --include='*.scss' --include='*.html' --include='*.metadata.json' --include='*/' --exclude='*' modules/ dist/modules/",
     "codecov": "codecov",
@@ -21,15 +22,15 @@
   "contributors": [],
   "license": "MIT",
   "devDependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
+    "@angular/animations": "^4.0.3",
+    "@angular/common": "^4.0.3",
+    "@angular/compiler": "^4.0.3",
+    "@angular/compiler-cli": "^4.0.3",
+    "@angular/core": "^4.0.3",
+    "@angular/forms": "^4.0.3",
+    "@angular/http": "^4.0.3",
+    "@angular/platform-browser": "^4.0.3",
+    "@angular/platform-browser-dynamic": "^4.0.3",
     "@types/es6-shim": "^0.31.32",
     "@types/jasmine": "^2.5.41",
     "@types/node": "^7.0.3",
@@ -65,13 +66,13 @@
     "ts-helpers": "1.1.2",
     "ts-node": "^3.0.2",
     "tslint": "^4.3.1",
-    "typescript": "2.2.1",
+    "typescript": "2.2.2",
     "universal-style-loader": "^0.16.3",
     "url-loader": "^0.5.7",
     "webpack": "^2.3.0",
-    "webpack-dev-server": "^2.4.0",
+    "webpack-dev-server": "https://registry.npmjs.org/webpack-dev-server/-/webpack-dev-server-2.4.2.tgz",
     "webpack-merge": "^4.1.0",
-    "zone.js": "0.8.5"
+    "zone.js": "0.8.7"
   },
   "dependencies": {
     "ng2-material-dropdown": "0.7.7"
