+++ tsconfig.aot.json
@@ -9,7 +9,7 @@
         "declaration": true,
         "sourceMap": true,
         "moduleResolution": "node",
-        "baseUrl": ".",
+        "baseUrl": "",
         "inlineSources": true,
         "skipLibCheck": true,
         "typeRoots": [ "./node_modules/@types" ],
@@ -33,10 +33,13 @@
     },
     "angularCompilerOptions": {
         "genDir": "factories",
-        "skipMetadataEmit": true
+        "strictMetadataEmit": true
     },
     "files": [
-        "./modules/index.ts"
+        "modules/index.ts"
+    ],
+    "include": [
+        "modules"
     ],
     "exclude": [
         "node_modules",
